+++ src/beat_link_trigger/triggers.clj
@@ -9,6 +9,7 @@
             [beat-link-trigger.menus :as menus]
             [beat-link-trigger.players :as players]
             [beat-link-trigger.auto-cache :as auto]
+            [beat-link-trigger.view-cache :as view-cache]
             [beat-link-trigger.prefs :as prefs]
             [beat-link-trigger.util :as util]
             [fipp.edn :as fipp]
@@ -987,6 +988,12 @@
                  :name "Auto-Attach Metadata Caches"
                  :key "menu M"))
 
+(def ^:private view-cache-action
+  "The menu action which allows the user to view the contents of a
+  metadata cache file."
+  (seesaw/action :handler (fn [e] (view-cache/choose-file @trigger-frame))
+                 :name "View Metadata Cache Contents"))
+
 (defn- midi-environment-changed
   "Called when CoreMidi4J reports a change to the MIDI environment, so we can update the menu of
   available MIDI outputs."
@@ -1304,7 +1311,7 @@
                        (.setPassive metadata-finder true))))
     (seesaw/menubar :items [(seesaw/menu :text "File"
                                          :items (concat [load-action save-action
-                                                         (seesaw/separator) auto-action
+                                                         (seesaw/separator) auto-action view-cache-action
                                                          (seesaw/separator) logs/logs-action]
                                                         menus/non-mac-actions))
                             (seesaw/menu :text "Triggers"